+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Minor: Allow overwriting webhooks or item lists on config import, rather than merging. (#595)
 - Minor: Allow specifying tag IDs to apply for new forum thread messages. (#599)
 - Minor: Add herbiboar support to kill count notifier. (#602)
 - Minor: Enable the leagues notifier again. (#597)